django_training_platform/requirements.txt
asgiref==3.3.1
attrs==20.3.0
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.2.1
defusedxml==0.7.0rc1
Django==3.1.4
django-debug-toolbar==3.2
django-filter==2.4.0
django-templated-mail==1.1.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
drf-spectacular==0.13.1
drf-yasg==1.20.0
factory-boy==3.2.0
Faker==6.1.1
idna==2.10
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==2.11.2
jsonschema==3.2.0
MarkupSafe==1.1.1
psycopg2-binary>=2.8
oauthlib==3.1.0
packaging==20.7
pluggy==0.13.1
py==1.10.0
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.2.2
pytest-django==4.1.0
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.4
PyYAML==5.4.1
requests==2.25.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.2
